Improved directories structure + removed useless servlets
