+++ demo/src/main/java/controllers/RegisterServlet.java
@@ -1,4 +1,4 @@
-package control;
+package controllers;
 
 import java.io.IOException;
 
@@ -6,7 +6,7 @@ import javax.servlet.ServletException;
 import javax.servlet.annotation.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import model.UserModel;
+import models.UserModel;
 
 import javax.servlet.http.HttpServlet;
 
@@ -17,6 +17,8 @@ public class RegisterServlet extends HttpServlet {
     
 	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
+
+		String page = "UserRegistered.jsp";
 
 		try {	
 			String user = request.getParameter("username");
