+++ demo/src/main/java/data/beans/TravelBean.java
@@ -1,4 +1,4 @@
-package model.beans;
+package data.beans;
 
 public class TravelBean {
