+++ demo/src/main/webapp/LogFail.jsp
@@ -1,5 +1,4 @@
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1" pageEncoding="ISO-8859-1"%>
-<%@ page import="model.beans.UserBean" %>
 <html>
     <head>
         <title>Logged</title>
